requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.3
attrs==22.2.0
billiard==4.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
curlify==2.2.1
exceptiongroup==1.1.3
Flask==2.3.3
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
kombu==5.3.2
MarkupSafe==2.1.3
maxminddb==2.4.0
more-itertools==10.1.0
multidict==6.0.4
Nasdaq-Data-Link==1.0.4
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
pdfminer.six==20221105
prompt-toolkit==3.0.39
proxybroker @ git+https://github.com/bluet/proxybroker2.git@19e57ba6333a965412eebe50c10aba9321815209
pyasn1==0.5.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.12.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
TradeRepublicApi @ git+https://github.com/Zarathustra2/TradeRepublicApi.git@519e7177407303a04b35ddbeda112a2138078a96
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zstandard==0.21.0
